Delete cookie from browser if session becomes empty
In a logout process this would allow us to user `$session->clear();` in
order to delete the session on the server side and now also delete the
cookie on the browser side.

While in most cases it's not critical to have leftover cookies on the browser
side, it is not a good practice. And it can become a real problem if use-case
where the cookie existence is used as a condition for some behavior, for
instance redirections via nginx configuration.

Signed-off-by: Adrien Crivelli <adrien.crivelli@gmail.com>